% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accidents}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numeric code of state of interest}

\item{year}{Year of interest}
}
\value{
A plot of the accidnts in a state for a given year.
}
\description{
\code{fars_map_state} Plots the coordinates of the accidents in a particular state for a specific year
If the numeric code for the state does not exist or there isn't a corresponding data file for the year of interest,
will not work
}
\examples{
fars_map_state(1, 2013)

}
