% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{path to csv file to be read.}
}
\value{
tibble with the FARS data for a particular year.
}
\description{
\code{fars_read} reads in the csv file containing the Fatality Analysis Reporting System data for a particular year
specified by the filename argument and returns it as a tibble.
Make sure there is a corresponding file for the year of interest
}
\examples{
fars_read("accident_2013.csv.bz2")

}
