% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read data for multiple years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Vector of years for which we want to read data.}
}
\value{
List of tibbles with two columns containing the month and year.
}
\description{
\code{fars_read_years} Reads in the data files for the years specified by the input. If any of the input years do not
have a corresponding data file, stops with error.
}
\examples{
fars_read_years(c(2013, 2014))

}
