% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize observations}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector of years for which we want to read data.}
}
\value{
A summary tibble which gives the no. of observations in each month for every year.
}
\description{
\code{fars_summarize_years} Summarizes the number of observations by month for the years specified in input
}
\examples{
fars_summarize_years(c(2013, 2014))

}
