% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make filename out of year specification}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or a string that can be coerced to integer specifying the year of interest.}
}
\value{
A string giving the name of the file that contains the year's data.
}
\description{
\code{make_filename} Given a year, creates the approriate name for the file that contains the data for that year
}
\examples{
make_filename(2013)
make_filename("2013")
}
