% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Shows in a map all the accidents on a given state.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A list of character strings or numeric values that provides the individual state}

\item{year}{A character string or a numeric that provides the year to be used in the file.}
}
\value{
This function plots in a map the locations where accidents occur for a given state.
}
\description{
Before plotting any accident on the map, this function contains two initial checks: first, by 
checking that the state number is valid and, second, by seeing if there were no accidents in that 
state. For the first type of check, the error message is "invalid STATE number: %d" with %d being
the state ID provided. The second type of check prints the message "no accidents to plot" if no
observations are present in the data frame for that state.
}
