% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reading CSV files with "fars_read" function and placing extracted data into data frame.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string that provides the location and name of the file to import.}
}
\value{
This function returns a data frame (tbl_df).
}
\description{
This is a simple function that reads a CSV file (after checking if it exists) and places
its content into a data frame without showing any messages to the user along the process.
}
