% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Showing table with month and year for each year in the list passed as an input.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of character strings or numeric values that provides the individual years to
be used in the file names}
}
\value{
This function simply returns or prints out tables with summaries split by year of the
   the number of observations per year and month
}
\description{
This function first joins the results provided by the function "fars_read_years", then groups the
data by the variables "year" and 'MONTH", followed by the grouping by the number of rows in each
group and finally splitting the year observations into columns.
}
