% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Printing "accident_%d.csv.bz2" where %d gets replaced by the year that gets passed as an input.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string or a numeric that provides the year to be used in the file name.
The first command within the function coerces the type of the variable year into an integer.}
}
\value{
NULL. This function simply prints out the name of the compressed file that should contain some 
   data.
}
\description{
This is a simple function that prints out the name of a compressed CSV file depending on the
year that gets passed to the function.
}
